pyproject.toml
[project]
name = "bustimes.org"
version = "0.1.0"
description = ""
authors = [{ name = "Josh Goodwin", email = "j@joshuagoodw.in" }]
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.12.1,<5",
    "boto3>=1.18.42,<2",
    "ciso8601>=2.2.0,<3",
    "dj-database-url>=3.0.0,<4",
    "Django>=5.2",
    "django-autoslug>=1.9.8,<2",
    "django-email-obfuscator>=0.1.5,<0.2",
    "django-filter~=25.1",
    "djangorestframework>=3.15",
    "django-simple-history>=3.0.0,<4",
    "django-sql-utils>=0.7.0,<0.8",
    "django-template-minifier>=1.1.0,<2",
    "gtfs-realtime-bindings~=2.0.0",
    "gunicorn>=25.1.0",
    "haversine>=2.5.1,<3",
    "huey>=2.4.3,<3",
    "psycopg[c]>=3.1.8,<4",
    "PyYAML>=6.0.0,<7",
    "redis>=7.3.0",
    "sentry-sdk[django]>=2.4.0,<3",
    "Shapely>=2.0.1,<3",
    "titlecase~=2.3",
    "uk-postcode-utils~=1.1",
    "webcolors~=25.10.0",
    "whitenoise>=6.2.0,<7",
    "django-cors-headers>=4.2.0,<5",
    "pygments>=2.15.1,<3",
    "django-turnstile>=0.1.0,<0.2",
    "jinja2>=3.1.6,<4",
    "requests>=2.32.3,<3",
    "numpy~=2.4.3",
    "tenacity>=9.0.0,<10",
    "lightningcss-cli>=1.29.2,<2",
    "websockets>=14.2",
    "django-multidb-router>=0.10",
    "folium>=0.20.0",
    "xmltodict>=1.0.2",
    "lxml>=6.0.0,<7",
    "django-http-compression>=1.2.0",
    "tqdm>=4.67.1",
    "polyline>=2.0.4",
    "django-timezone-field>=7.2.1",
    "django-imagekit>=6.1.0",
    "django-storages>=1.14.6",
    "transxchange-parser",
]

[dependency-groups]
test = [
    "govuk-bank-holidays>=0.15",
    "asgiref>=3.4.1,<4",
    "gtfs-kit>=11.0.0",
    "requests-toolbelt>=1.0.0,<2",
    "time-machine~=3.2.0",
    "vcrpy~=8.1.0",
    "fakeredis>=2.12.1,<3",
    "tblib>=3.0.0,<4",
    "coverage>=7.6.1,<8",
]
dev = [
    "django-debug-toolbar>=6.0.0,<7",
    "django-debug-toolbar-force>=0.2,<0.3",
    "django-debug-toolbar-template-profiler>=2.1.0,<3",
    "django-watchfiles>=1.4.0",
    "ipdb>=0.13.9,<0.14",
    "ipython>=9.0.2,<10",
    "uritemplate>=4.1.1,<5",
]

[tool.uv]
package = false

[tool.uv.sources]
transxchange-parser = { git = "https://github.com/bustimes/transxchange-parser" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.coverage.run]
relative_files = true
source = ["."]
omit =  ["*/migrations/*"]

[tool.ruff]
unsafe-fixes = true

[tool.ruff.lint]
extend-select = ["SIM117"]
